added ternary logic to show different login buttons depending on state
