+++ rightmove/src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import './App.css';
-import Nav from "./components/Nav";
+import Nav from "./components/Nav/Nav";
 
 function App() {
   return (
     <Router>
       <>
-        <Nav />
+        <Nav name="Steve"/>
       </>
     </Router>
   );
