+++ rightmove/src/components/Col/Col.js
