+++ rightmove/src/components/Nav/style.css
@@ -1,0 +1,18 @@
+nav{
+  background: none;
+}
+
+a.brand-logo{
+  color: black;
+}
+
+nav ul a{
+  color: black;
+}
+
+.userName{
+  font-size: 100%;
+  color:black;
+
+}
+
